decoupled controllers from Spring MVC
