+++ server/pxf-service/src/main/java/org/greenplum/pxf/service/controller/ReadService.java
@@ -1,7 +1,6 @@
 package org.greenplum.pxf.service.controller;
 
 import org.greenplum.pxf.api.model.RequestContext;
-import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
 
 /**
  * Service that reads data from external systems.
@@ -14,5 +13,5 @@ public interface ReadService {
      * @param context request context
      * @return read response instance
      */
-    StreamingResponseBody getReadResponse(RequestContext context);
+    ReadResponse getReadResponse(RequestContext context);
 }
